fix: Rename CreateScheduleDialog, CreateScheduleContract, UpdateAndDeleteScheduleDialog, UpdateAndDeleteScheduleContract
