+++ app/src/main/AndroidManifest.xml
@@ -17,14 +17,15 @@
         tools:targetApi="31">
 
         <activity
-            android:name=".ui.bottomsheet.BottomSheetActivity"
-            android:exported="false"
-            android:theme="@style/Theme.AppCompat.Light.NoActionBar"/>
-        <activity
             android:name=".MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.Together_watch">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
         </activity>
 
         <activity
