+++ app/src/main/java/com/example/together_watch/schedule/CreateScheduleModel.kt
@@ -1,0 +1,11 @@
+package com.example.together_watch.schedule
+
+import com.example.together_watch.ui.schedule.Schedule
+import com.google.firebase.Firebase
+import com.google.firebase.auth.auth
+import com.google.firebase.firestore.FirebaseFirestore
+
+class CreateScheduleModel : CreateScheduleContract.Model {
+    override fun savePersonalSchedule(schedule: Schedule) {
+    }
+}
