+++ app/src/main/java/com/example/together_watch/schedule/UpdateAndDeleteScheduleDialog.kt
@@ -1,0 +1,4 @@
+package com.example.together_watch.schedule
+
+class UpdateAndDeleteScheduleDialog {
+}
