--- app/src/main/java/com/example/together_watch/ui/bottomsheet/BottomSheetUpdateAndDeleteDialog.kt
@@ -1,4 +1,0 @@
-package com.example.together_watch.ui.bottomsheet
-
-class BottomSheetUpdateAndDeleteDialog {
-}
